Splitting up delta_generator, part 2 (#1776)
* Use the typed element proto, not Element

* Extract out st.markdown and st.button

These two serve as test cases for splitting up; one element and one widget

* Move widget helpers into utils to break circular dependency

SQUASH

* Extract out st.balloons, latex, text

* Extract out all markdown-like commands

* Fixup code by testing against an actual streamlit app

* Move st.text out of markdown

* Extract alerts, and consolidate MarkdownMixin

* Fixup

* Fix balloons

* Extract st.help

* Extract st.exception

* Support @_with_element st.commands during transition

* Fix up python unit tests

* Include the type as an input to the widget id hash

* Overwrite the __module__ of mixined commands

* Ignore mypy error from missing in mixins

* Clean up mypy errors

* dg._enqueue should always be returned from st.methods

* Drop noqa

* Replace Foo_pb2.Foo with FooProto

* Extract out st.dataframe

* Extract out altair chart types

* Extract out the actual st.altair_chart

* Extract VegaLite, Bokeh, Graphviz, Plotly charts

* dg._enqueue should always return, redux

* Fixup bokeh

* Fix add_rows breakage

* mypy cleanup redux

* Extract table, deck_gl, pydeck, map

* Replace foo_pb2 with FooProto, redux

* Try moving bokeh into dev-packages to fix RTD error

* Set bokeh as required package

* Fix PydeckProto import

* Only import bokeh once st.bokeh_chart is called
