+++ lib/streamlit/delta_generator.py
@@ -35,8 +35,6 @@ from streamlit.errors import NoSessionContext
 from streamlit.file_util import get_encoded_file_data
 from streamlit.js_number import JSNumber
 from streamlit.js_number import JSNumberBoundsException
-from streamlit.proto import Alert_pb2
-from streamlit.proto import Balloons_pb2
 from streamlit.proto import BlockPath_pb2
 from streamlit.proto import ForwardMsg_pb2
 from streamlit.proto.Element_pb2 import Element
@@ -55,6 +53,15 @@ from streamlit.elements.alert import AlertMixin
 from streamlit.elements.json import JsonMixin
 from streamlit.elements.doc_string import HelpMixin
 from streamlit.elements.exception_proto import ExceptionMixin
+from streamlit.elements.data_frame_proto import DataFrameMixin
+from streamlit.elements.altair import AltairMixin
+from streamlit.elements.bokeh_chart import BokehMixin
+from streamlit.elements.graphviz_chart import GraphvizMixin
+from streamlit.elements.plotly_chart import PlotlyMixin
+from streamlit.elements.vega_lite import VegaLiteMixin
+from streamlit.elements.deck_gl import DeckGlMixin
+from streamlit.elements.deck_gl_json_chart import PydeckMixin
+from streamlit.elements.map import MapMixin
 
 LOGGER = get_logger(__name__)
 
@@ -199,13 +206,22 @@ for more information.
 
 class DeltaGenerator(
     AlertMixin,
+    AltairMixin,
     BalloonsMixin,
+    BokehMixin,
     ButtonMixin,
+    DataFrameMixin,
+    DeckGlMixin,
     ExceptionMixin,
+    GraphvizMixin,
     HelpMixin,
     MarkdownMixin,
+    MapMixin,
+    PlotlyMixin,
+    PydeckMixin,
     JsonMixin,
     TextMixin,
+    VegaLiteMixin,
 ):
     """Creator of Delta protobuf messages.
 
@@ -350,11 +366,17 @@ class DeltaGenerator(
         # Warn if we're called from within an @st.cache function
         caching.maybe_show_cached_st_function_warning(self, delta_type)
 
-        # TODO: DELTAS_TYPES_THAT_MELT_DATAFRAMES mixins should fill last_index
+        # Some elements have a method.__name__ != delta_type in proto.
+        # This really matters for line_chart, bar_chart & area_chart,
+        # since add_rows() relies on method.__name__ == delta_type
+        # TODO: Fix for all elements (or the cache warning above will be wrong)
+        proto_type = delta_type
+        if proto_type in DELTAS_TYPES_THAT_MELT_DATAFRAMES:
+            proto_type = "vega_lite_chart"
 
         # Copy the marshalled proto into the overall msg proto
         msg = ForwardMsg_pb2.ForwardMsg()
-        msg_el_proto = getattr(msg.delta.new_element, delta_type)
+        msg_el_proto = getattr(msg.delta.new_element, proto_type)
         msg_el_proto.CopyFrom(element_proto)
 
         # Only enqueue message and fill in metadata if there's a container.
@@ -477,551 +499,6 @@ class DeltaGenerator(
 
         return block_dg
 
-    def dataframe(self, data=None, width=None, height=None):
-        """Display a dataframe as an interactive table.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict,
-            or None
-            The data to display.
-
-            If 'data' is a pandas.Styler, it will be used to style its
-            underyling DataFrame. Streamlit supports custom cell
-            values and colors. (It does not support some of the more exotic
-            pandas styling features, like bar charts, hovering, and captions.)
-            Styler support is experimental!
-        width : int or None
-            Desired width of the UI element expressed in pixels. If None, a
-            default width based on the page width is used.
-        height : int or None
-            Desired height of the UI element expressed in pixels. If None, a
-            default height is used.
-
-        Examples
-        --------
-        >>> df = pd.DataFrame(
-        ...    np.random.randn(50, 20),
-        ...    columns=('col %d' % i for i in range(20)))
-        ...
-        >>> st.dataframe(df)  # Same as st.write(df)
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=165mJbzWdAC8Duf8a4tjyQ
-           height: 330px
-
-        >>> st.dataframe(df, 200, 100)
-
-        You can also pass a Pandas Styler object to change the style of
-        the rendered DataFrame:
-
-        >>> df = pd.DataFrame(
-        ...    np.random.randn(10, 20),
-        ...    columns=('col %d' % i for i in range(20)))
-        ...
-        >>> st.dataframe(df.style.highlight_max(axis=0))
-
-        .. output::
-           https://share.streamlit.io/0.29.0-dV1Y/index.html?id=Hb6UymSNuZDzojUNybzPby
-           height: 285px
-
-        """
-        import streamlit.elements.data_frame_proto as data_frame_proto
-
-        def set_data_frame(delta):
-            data_frame_proto.marshall_data_frame(data, delta.data_frame)
-
-        return self._enqueue_new_element_delta(
-            set_data_frame, "dataframe", element_width=width, element_height=height
-        )
-
-    @_with_element
-    def line_chart(
-        self, element, data=None, width=0, height=0, use_container_width=True
-    ):
-        """Display a line chart.
-
-        This is just syntax-sugar around st.altair_chart. The main difference
-        is this command uses the data's own column and indices to figure out
-        the chart's spec. As a result this is easier to use for many "just plot
-        this" scenarios, while being less customizable.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict
-            or None
-            Data to be plotted.
-
-        width : int
-            The chart width in pixels. If 0, selects the width automatically.
-
-        height : int
-            The chart width in pixels. If 0, selects the height automatically.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over the width argument.
-
-        Example
-        -------
-        >>> chart_data = pd.DataFrame(
-        ...     np.random.randn(20, 3),
-        ...     columns=['a', 'b', 'c'])
-        ...
-        >>> st.line_chart(chart_data)
-
-        .. output::
-           https://share.streamlit.io/0.50.0-td2L/index.html?id=BdxXG3MmrVBfJyqS2R2ki8
-           height: 220px
-
-        """
-
-        import streamlit.elements.altair as altair
-
-        chart = altair.generate_chart("line", data, width, height)
-        altair.marshall(element.vega_lite_chart, chart, use_container_width)
-
-    @_with_element
-    def area_chart(
-        self, element, data=None, width=0, height=0, use_container_width=True
-    ):
-        """Display a area chart.
-
-        This is just syntax-sugar around st.altair_chart. The main difference
-        is this command uses the data's own column and indices to figure out
-        the chart's spec. As a result this is easier to use for many "just plot
-        this" scenarios, while being less customizable.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, or dict
-            Data to be plotted.
-
-        width : int
-            The chart width in pixels. If 0, selects the width automatically.
-
-        height : int
-            The chart width in pixels. If 0, selects the height automatically.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over the width argument.
-
-        Example
-        -------
-        >>> chart_data = pd.DataFrame(
-        ...     np.random.randn(20, 3),
-        ...     columns=['a', 'b', 'c'])
-        ...
-        >>> st.area_chart(chart_data)
-
-        .. output::
-           https://share.streamlit.io/0.50.0-td2L/index.html?id=Pp65STuFj65cJRDfhGh4Jt
-           height: 220px
-
-        """
-        import streamlit.elements.altair as altair
-
-        chart = altair.generate_chart("area", data, width, height)
-        altair.marshall(element.vega_lite_chart, chart, use_container_width)
-
-    @_with_element
-    def bar_chart(
-        self, element, data=None, width=0, height=0, use_container_width=True
-    ):
-        """Display a bar chart.
-
-        This is just syntax-sugar around st.altair_chart. The main difference
-        is this command uses the data's own column and indices to figure out
-        the chart's spec. As a result this is easier to use for many "just plot
-        this" scenarios, while being less customizable.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, or dict
-            Data to be plotted.
-
-        width : int
-            The chart width in pixels. If 0, selects the width automatically.
-
-        height : int
-            The chart width in pixels. If 0, selects the height automatically.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over the width argument.
-
-        Example
-        -------
-        >>> chart_data = pd.DataFrame(
-        ...     np.random.randn(50, 3),
-        ...     columns=["a", "b", "c"])
-        ...
-        >>> st.bar_chart(chart_data)
-
-        .. output::
-           https://share.streamlit.io/0.50.0-td2L/index.html?id=5U5bjR2b3jFwnJdDfSvuRk
-           height: 220px
-
-        """
-        import streamlit.elements.altair as altair
-
-        chart = altair.generate_chart("bar", data, width, height)
-        altair.marshall(element.vega_lite_chart, chart, use_container_width)
-
-    @_with_element
-    def vega_lite_chart(
-        self,
-        element,
-        data=None,
-        spec=None,
-        width=0,
-        use_container_width=False,
-        **kwargs,
-    ):
-        """Display a chart using the Vega-Lite library.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict,
-            or None
-            Either the data to be plotted or a Vega-Lite spec containing the
-            data (which more closely follows the Vega-Lite API).
-
-        spec : dict or None
-            The Vega-Lite spec for the chart. If the spec was already passed in
-            the previous argument, this must be set to None. See
-            https://vega.github.io/vega-lite/docs/ for more info.
-
-        width : number
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the width directly in the Vega-Lite
-            spec. Please refer to the Vega-Lite documentation for details.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over Vega-Lite's native `width` value.
-
-        **kwargs : any
-            Same as spec, but as keywords.
-
-        Example
-        -------
-
-        >>> import pandas as pd
-        >>> import numpy as np
-        >>>
-        >>> df = pd.DataFrame(
-        ...     np.random.randn(200, 3),
-        ...     columns=['a', 'b', 'c'])
-        >>>
-        >>> st.vega_lite_chart(df, {
-        ...     'mark': {'type': 'circle', 'tooltip': True},
-        ...     'encoding': {
-        ...         'x': {'field': 'a', 'type': 'quantitative'},
-        ...         'y': {'field': 'b', 'type': 'quantitative'},
-        ...         'size': {'field': 'c', 'type': 'quantitative'},
-        ...         'color': {'field': 'c', 'type': 'quantitative'},
-        ...     },
-        ... })
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=8jmmXR8iKoZGV4kXaKGYV5
-           height: 200px
-
-        Examples of Vega-Lite usage without Streamlit can be found at
-        https://vega.github.io/vega-lite/examples/. Most of those can be easily
-        translated to the syntax shown above.
-
-        """
-        import streamlit.elements.vega_lite as vega_lite
-
-        if width != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` argument in `st.vega_lite_chart` is deprecated and will be removed on 2020-03-04. To set the width, you should instead use Vega-Lite's native `width` argument as described at https://vega.github.io/vega-lite/docs/size.html"
-            )
-
-        vega_lite.marshall(
-            element.vega_lite_chart,
-            data,
-            spec,
-            use_container_width=use_container_width,
-            **kwargs,
-        )
-
-    @_with_element
-    def altair_chart(self, element, altair_chart, width=0, use_container_width=False):
-        """Display a chart using the Altair library.
-
-        Parameters
-        ----------
-        altair_chart : altair.vegalite.v2.api.Chart
-            The Altair chart object to display.
-
-        width : number
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the width directly in the Altair
-            spec. Please refer to the Altair documentation for details.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over Altair's native `width` value.
-
-        Example
-        -------
-
-        >>> import pandas as pd
-        >>> import numpy as np
-        >>> import altair as alt
-        >>>
-        >>> df = pd.DataFrame(
-        ...     np.random.randn(200, 3),
-        ...     columns=['a', 'b', 'c'])
-        ...
-        >>> c = alt.Chart(df).mark_circle().encode(
-        ...     x='a', y='b', size='c', color='c', tooltip=['a', 'b', 'c'])
-        >>>
-        >>> st.altair_chart(c, use_container_width=True)
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=8jmmXR8iKoZGV4kXaKGYV5
-           height: 200px
-
-        Examples of Altair charts can be found at
-        https://altair-viz.github.io/gallery/.
-
-        """
-        import streamlit.elements.altair as altair
-
-        if width != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` argument in `st.vega_lite_chart` is deprecated and will be removed on 2020-03-04. To set the width, you should instead use altair's native `width` argument as described at https://altair-viz.github.io/user_guide/generated/toplevel/altair.Chart.html"
-            )
-
-        altair.marshall(
-            element.vega_lite_chart,
-            altair_chart,
-            use_container_width=use_container_width,
-        )
-
-    @_with_element
-    def graphviz_chart(
-        self, element, figure_or_dot, width=0, height=0, use_container_width=False
-    ):
-        """Display a graph using the dagre-d3 library.
-
-        Parameters
-        ----------
-        figure_or_dot : graphviz.dot.Graph, graphviz.dot.Digraph, str
-            The Graphlib graph object or dot string to display
-
-        width : number
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the width directly in the Graphviz
-            spec. Please refer to the Graphviz documentation for details.
-
-        height : number
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the height directly in the Graphviz
-            spec. Please refer to the Graphviz documentation for details.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over the figure's native `width` value.
-
-        Example
-        -------
-
-        >>> import streamlit as st
-        >>> import graphviz as graphviz
-        >>>
-        >>> # Create a graphlib graph object
-        >>> graph = graphviz.Digraph()
-        >>> graph.edge('run', 'intr')
-        >>> graph.edge('intr', 'runbl')
-        >>> graph.edge('runbl', 'run')
-        >>> graph.edge('run', 'kernel')
-        >>> graph.edge('kernel', 'zombie')
-        >>> graph.edge('kernel', 'sleep')
-        >>> graph.edge('kernel', 'runmem')
-        >>> graph.edge('sleep', 'swap')
-        >>> graph.edge('swap', 'runswap')
-        >>> graph.edge('runswap', 'new')
-        >>> graph.edge('runswap', 'runmem')
-        >>> graph.edge('new', 'runmem')
-        >>> graph.edge('sleep', 'runmem')
-        >>>
-        >>> st.graphviz_chart(graph)
-
-        Or you can render the chart from the graph using GraphViz's Dot
-        language:
-
-        >>> st.graphviz_chart('''
-            digraph {
-                run -> intr
-                intr -> runbl
-                runbl -> run
-                run -> kernel
-                kernel -> zombie
-                kernel -> sleep
-                kernel -> runmem
-                sleep -> swap
-                swap -> runswap
-                runswap -> new
-                runswap -> runmem
-                new -> runmem
-                sleep -> runmem
-            }
-        ''')
-
-        .. output::
-           https://share.streamlit.io/0.56.0-xTAd/index.html?id=GBn3GXZie5K1kXuBKe4yQL
-           height: 400px
-
-        """
-        import streamlit.elements.graphviz_chart as graphviz_chart
-
-        if width != 0 and height != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` and `height` arguments in `st.graphviz` are deprecated and will be removed on 2020-03-04"
-            )
-        elif width != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` argument in `st.graphviz` is deprecated and will be removed on 2020-03-04"
-            )
-        elif height != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `height` argument in `st.graphviz` is deprecated and will be removed on 2020-03-04"
-            )
-
-        graphviz_chart.marshall(
-            element.graphviz_chart, figure_or_dot, use_container_width
-        )
-
-    @_with_element
-    def plotly_chart(
-        self,
-        element,
-        figure_or_data,
-        width=0,
-        height=0,
-        use_container_width=False,
-        sharing="streamlit",
-        **kwargs,
-    ):
-        """Display an interactive Plotly chart.
-
-        Plotly is a charting library for Python. The arguments to this function
-        closely follow the ones for Plotly's `plot()` function. You can find
-        more about Plotly at https://plot.ly/python.
-
-        Parameters
-        ----------
-        figure_or_data : plotly.graph_objs.Figure, plotly.graph_objs.Data,
-            dict/list of plotly.graph_objs.Figure/Data
-
-            See https://plot.ly/python/ for examples of graph descriptions.
-
-        width : int
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the width directly in the figure.
-            Please refer to the Plotly documentation for details.
-
-        height : int
-            Deprecated. If != 0 (default), will show an alert.
-            From now on you should set the height directly in the figure.
-            Please refer to the Plotly documentation for details.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over the figure's native `width` value.
-
-        sharing : {'streamlit', 'private', 'secret', 'public'}
-            Use 'streamlit' to insert the plot and all its dependencies
-            directly in the Streamlit app, which means it works offline too.
-            This is the default.
-            Use any other sharing mode to send the app to Plotly's servers,
-            and embed the result into the Streamlit app. See
-            https://plot.ly/python/privacy/ for more. Note that these sharing
-            modes require a Plotly account.
-
-        **kwargs
-            Any argument accepted by Plotly's `plot()` function.
-
-
-        To show Plotly charts in Streamlit, just call `st.plotly_chart`
-        wherever you would call Plotly's `py.plot` or `py.iplot`.
-
-        Example
-        -------
-
-        The example below comes straight from the examples at
-        https://plot.ly/python:
-
-        >>> import streamlit as st
-        >>> import plotly.figure_factory as ff
-        >>> import numpy as np
-        >>>
-        >>> # Add histogram data
-        >>> x1 = np.random.randn(200) - 2
-        >>> x2 = np.random.randn(200)
-        >>> x3 = np.random.randn(200) + 2
-        >>>
-        >>> # Group data together
-        >>> hist_data = [x1, x2, x3]
-        >>>
-        >>> group_labels = ['Group 1', 'Group 2', 'Group 3']
-        >>>
-        >>> # Create distplot with custom bin_size
-        >>> fig = ff.create_distplot(
-        ...         hist_data, group_labels, bin_size=[.1, .25, .5])
-        >>>
-        >>> # Plot!
-        >>> st.plotly_chart(fig, use_container_width=True)
-
-        .. output::
-           https://share.streamlit.io/0.56.0-xTAd/index.html?id=TuP96xX8JnsoQeUGAPjkGQ
-           height: 400px
-
-        """
-        # NOTE: "figure_or_data" is the name used in Plotly's .plot() method
-        # for their main parameter. I don't like the name, but it's best to
-        # keep it in sync with what Plotly calls it.
-        import streamlit.elements.plotly_chart as plotly_chart
-
-        if width != 0 and height != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` and `height` arguments in `st.plotly_chart` are deprecated and will be removed on 2020-03-04. To set these values, you should instead use Plotly's native arguments as described at https://plot.ly/python/setting-graph-size/"
-            )
-        elif width != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `width` argument in `st.plotly_chart` is deprecated and will be removed on 2020-03-04. To set the width, you should instead use Plotly's native `width` argument as described at https://plot.ly/python/setting-graph-size/"
-            )
-        elif height != 0:
-            import streamlit as st
-
-            st.warning(
-                "The `height` argument in `st.plotly_chart` is deprecated and will be removed on 2020-03-04. To set the height, you should instead use Plotly's native `height` argument as described at https://plot.ly/python/setting-graph-size/"
-            )
-
-        plotly_chart.marshall(
-            element.plotly_chart, figure_or_data, use_container_width, sharing, **kwargs
-        )
-
     @_with_element
     def pyplot(self, element, fig=None, clear_figure=None, **kwargs):
         """Display a matplotlib.pyplot figure.
@@ -1073,52 +550,6 @@ class DeltaGenerator(
         import streamlit.elements.pyplot as pyplot
 
         pyplot.marshall(self._get_coordinates, element, fig, clear_figure, **kwargs)
-
-    @_with_element
-    def bokeh_chart(self, element, figure, use_container_width=False):
-        """Display an interactive Bokeh chart.
-
-        Bokeh is a charting library for Python. The arguments to this function
-        closely follow the ones for Bokeh's `show` function. You can find
-        more about Bokeh at https://bokeh.pydata.org.
-
-        Parameters
-        ----------
-        figure : bokeh.plotting.figure.Figure
-            A Bokeh figure to plot.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over Bokeh's native `width` value.
-
-        To show Bokeh charts in Streamlit, just call `st.bokeh_chart`
-        wherever you would call Bokeh's `show`.
-
-        Example
-        -------
-        >>> import streamlit as st
-        >>> from bokeh.plotting import figure
-        >>>
-        >>> x = [1, 2, 3, 4, 5]
-        >>> y = [6, 7, 2, 4, 5]
-        >>>
-        >>> p = figure(
-        ...     title='simple line example',
-        ...     x_axis_label='x',
-        ...     y_axis_label='y')
-        ...
-        >>> p.line(x, y, legend='Trend', line_width=2)
-        >>>
-        >>> st.bokeh_chart(p, use_container_width=True)
-
-        .. output::
-           https://share.streamlit.io/0.56.0-xTAd/index.html?id=Fdhg51uMbGMLRRxXV6ubzp
-           height: 600px
-
-        """
-        import streamlit.elements.bokeh_chart as bokeh_chart
-
-        bokeh_chart.marshall(element.bokeh_chart, figure, use_container_width)
 
     @_with_element
     def image(
@@ -2666,282 +2097,13 @@ class DeltaGenerator(
         # The protobuf needs something to be set
         element.empty.unused = True
 
-    @_with_element
-    def map(self, element, data=None, zoom=None, use_container_width=True):
-        """Display a map with points on it.
-
-        This is a wrapper around st.pydeck_chart to quickly create scatterplot
-        charts on top of a map, with auto-centering and auto-zoom.
-
-        When using this command, we advise all users to use a personal Mapbox
-        token. This ensures the map tiles used in this chart are more
-        robust. You can do this with the mapbox.token config option.
-
-        To get a token for yourself, create an account at
-        https://mapbox.com. It's free! (for moderate usage levels) See
-        https://docs.streamlit.io/en/latest/cli.html#view-all-config-options for more
-        info on how to set config options.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict,
-            or None
-            The data to be plotted. Must have columns called 'lat', 'lon',
-            'latitude', or 'longitude'.
-        zoom : int
-            Zoom level as specified in
-            https://wiki.openstreetmap.org/wiki/Zoom_levels
-
-        Example
-        -------
-        >>> import pandas as pd
-        >>> import numpy as np
-        >>>
-        >>> df = pd.DataFrame(
-        ...     np.random.randn(1000, 2) / [50, 50] + [37.76, -122.4],
-        ...     columns=['lat', 'lon'])
-        >>>
-        >>> st.map(df)
-
-        .. output::
-           https://share.streamlit.io/0.53.0-SULT/index.html?id=9gTiomqPEbvHY2huTLoQtH
-           height: 600px
-
-        """
-        import streamlit.elements.map as streamlit_map
-
-        element.deck_gl_json_chart.json = streamlit_map.to_deckgl_json(data, zoom)
-        element.deck_gl_json_chart.use_container_width = use_container_width
-
-    @_with_element
-    def deck_gl_chart(self, element, spec=None, use_container_width=False, **kwargs):
-        """Draw a map chart using the Deck.GL library.
-
-        This API closely follows Deck.GL's JavaScript API
-        (https://deck.gl/#/documentation), with a few small adaptations and
-        some syntax sugar.
-
-        When using this command, we advise all users to use a personal Mapbox
-        token. This ensures the map tiles used in this chart are more
-        robust. You can do this with the mapbox.token config option.
-
-        To get a token for yourself, create an account at
-        https://mapbox.com. It's free! (for moderate usage levels) See
-        https://docs.streamlit.io/en/latest/cli.html#view-all-config-options for more
-        info on how to set config options.
-
-        Parameters
-        ----------
-
-        spec : dict
-            Keys in this dict can be:
-
-            - Anything accepted by Deck.GL's top level element, such as
-              "viewport", "height", "width".
-
-            - "layers": a list of dicts containing information to build a new
-              Deck.GL layer in the map. Each layer accepts the following keys:
-
-                - "data" : DataFrame
-                  The data for the current layer.
-
-                - "type" : str
-                  One of the Deck.GL layer types that are currently supported
-                  by Streamlit: ArcLayer, GridLayer, HexagonLayer, LineLayer,
-                  PointCloudLayer, ScatterplotLayer, ScreenGridLayer,
-                  TextLayer.
-
-                - Plus anything accepted by that layer type. The exact keys that
-                  are accepted depend on the "type" field, above. For example, for
-                  ScatterplotLayer you can set fields like "opacity", "filled",
-                  "stroked", and so on.
-
-                  In addition, Deck.GL"s documentation for ScatterplotLayer
-                  shows you can use a "getRadius" field to individually set
-                  the radius of each circle in the plot. So here you would
-                  set "getRadius": "my_column" where "my_column" is the name
-                  of the column containing the radius data.
-
-                  For things like "getPosition", which expect an array rather
-                  than a scalar value, we provide alternates that make the
-                  API simpler to use with dataframes:
-
-                  - Instead of "getPosition" : use "getLatitude" and
-                    "getLongitude".
-                  - Instead of "getSourcePosition" : use "getLatitude" and
-                    "getLongitude".
-                  - Instead of "getTargetPosition" : use "getTargetLatitude"
-                    and "getTargetLongitude".
-                  - Instead of "getColor" : use "getColorR", "getColorG",
-                    "getColorB", and (optionally) "getColorA", for red,
-                    green, blue and alpha.
-                  - Instead of "getSourceColor" : use the same as above.
-                  - Instead of "getTargetColor" : use "getTargetColorR", etc.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over the figure's native `width` value.
-
-        **kwargs : any
-            Same as spec, but as keywords. Keys are "unflattened" at the
-            underscore characters. For example, foo_bar_baz=123 becomes
-            foo={'bar': {'bar': 123}}.
-
-        Example
-        -------
-        >>> st.deck_gl_chart(
-        ...     viewport={
-        ...         'latitude': 37.76,
-        ...         'longitude': -122.4,
-        ...         'zoom': 11,
-        ...         'pitch': 50,
-        ...     },
-        ...     layers=[{
-        ...         'type': 'HexagonLayer',
-        ...         'data': df,
-        ...         'radius': 200,
-        ...         'elevationScale': 4,
-        ...         'elevationRange': [0, 1000],
-        ...         'pickable': True,
-        ...         'extruded': True,
-        ...     }, {
-        ...         'type': 'ScatterplotLayer',
-        ...         'data': df,
-        ...     }])
-        ...
-
-        .. output::
-           https://share.streamlit.io/0.50.0-td2L/index.html?id=3GfRygWqxuqB5UitZLjz9i
-           height: 530px
-
-        """
-
-        suppress_deprecation_warning = config.get_option(
-            "global.suppressDeprecationWarnings"
-        )
-        if not suppress_deprecation_warning:
-            import streamlit as st
-
-            st.warning(
-                """
-                The `deck_gl_chart` widget is deprecated and will be removed on
-                2020-05-01. To render a map, you should use `st.pydeck_chart` widget.
-            """
-            )
-
-        import streamlit.elements.deck_gl as deck_gl
-
-        deck_gl.marshall(element.deck_gl_chart, spec, use_container_width, **kwargs)
-
-    @_with_element
-    def pydeck_chart(self, element, pydeck_obj=None, use_container_width=False):
-        """Draw a chart using the PyDeck library.
-
-        This supports 3D maps, point clouds, and more! More info about PyDeck
-        at https://deckgl.readthedocs.io/en/latest/.
-
-        These docs are also quite useful:
-
-        - DeckGL docs: https://github.com/uber/deck.gl/tree/master/docs
-        - DeckGL JSON docs: https://github.com/uber/deck.gl/tree/master/modules/json
-
-        When using this command, we advise all users to use a personal Mapbox
-        token. This ensures the map tiles used in this chart are more
-        robust. You can do this with the mapbox.token config option.
-
-        To get a token for yourself, create an account at
-        https://mapbox.com. It's free! (for moderate usage levels) See
-        https://docs.streamlit.io/en/latest/cli.html#view-all-config-options for more
-        info on how to set config options.
-
-        Parameters
-        ----------
-        spec: pydeck.Deck or None
-            Object specifying the PyDeck chart to draw.
-
-        Example
-        -------
-        Here's a chart using a HexagonLayer and a ScatterplotLayer on top of
-        the light map style:
-
-        >>> df = pd.DataFrame(
-        ...    np.random.randn(1000, 2) / [50, 50] + [37.76, -122.4],
-        ...    columns=['lat', 'lon'])
-        >>>
-        >>> st.pydeck_chart(pdk.Deck(
-        ...     map_style='mapbox://styles/mapbox/light-v9',
-        ...     initial_view_state=pdk.ViewState(
-        ...         latitude=37.76,
-        ...         longitude=-122.4,
-        ...         zoom=11,
-        ...         pitch=50,
-        ...     ),
-        ...     layers=[
-        ...         pdk.Layer(
-        ...            'HexagonLayer',
-        ...            data=df,
-        ...            get_position='[lon, lat]',
-        ...            radius=200,
-        ...            elevation_scale=4,
-        ...            elevation_range=[0, 1000],
-        ...            pickable=True,
-        ...            extruded=True,
-        ...         ),
-        ...         pdk.Layer(
-        ...             'ScatterplotLayer',
-        ...             data=df,
-        ...             get_position='[lon, lat]',
-        ...             get_color='[200, 30, 0, 160]',
-        ...             get_radius=200,
-        ...         ),
-        ...     ],
-        ... ))
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=ASTdExBpJ1WxbGceneKN1i
-           height: 530px
-
-        """
-        import streamlit.elements.deck_gl_json_chart as deck_gl_json_chart
-
-        deck_gl_json_chart.marshall(element, pydeck_obj, use_container_width)
-
-    @_with_element
-    def table(self, element, data=None):
-        """Display a static table.
-
-        This differs from `st.dataframe` in that the table in this case is
-        static: its entire contents are just laid out directly on the page.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict,
-            or None
-            The table data.
-
-        Example
-        -------
-        >>> df = pd.DataFrame(
-        ...    np.random.randn(10, 5),
-        ...    columns=('col %d' % i for i in range(5)))
-        ...
-        >>> st.table(df)
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=KfZvDMprL4JFKXbpjD3fpq
-           height: 480px
-
-        """
-        import streamlit.elements.data_frame_proto as data_frame_proto
-
-        data_frame_proto.marshall_data_frame(data, element.table)
-
     def add_rows(self, data=None, **kwargs):
         """Concatenate a dataframe to the bottom of the current one.
 
         Parameters
         ----------
-        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict, or None
+        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict,
+        or None
             Table to concat. Optional.
 
         **kwargs : pandas.DataFrame, numpy.ndarray, Iterable, dict, or None
