+++ lib/streamlit/elements/altair.py
@@ -17,9 +17,204 @@
 from datetime import date
 
 from streamlit import type_util
+from streamlit.proto.VegaLiteChart_pb2 import VegaLiteChart as VegaLiteChartProto
 import streamlit.elements.vega_lite as vega_lite
 import altair as alt
 import pandas as pd
+
+from .utils import last_index_for_melted_dataframes
+
+
+class AltairMixin:
+    def line_chart(dg, data=None, width=0, height=0, use_container_width=True):
+        """Display a line chart.
+
+        This is just syntax-sugar around st.altair_chart. The main difference
+        is this command uses the data's own column and indices to figure out
+        the chart's spec. As a result this is easier to use for many "just plot
+        this" scenarios, while being less customizable.
+
+        Parameters
+        ----------
+        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, dict
+            or None
+            Data to be plotted.
+
+        width : int
+            The chart width in pixels. If 0, selects the width automatically.
+
+        height : int
+            The chart width in pixels. If 0, selects the height automatically.
+
+        use_container_width : bool
+            If True, set the chart width to the column width. This takes
+            precedence over the width argument.
+
+        Example
+        -------
+        >>> chart_data = pd.DataFrame(
+        ...     np.random.randn(20, 3),
+        ...     columns=['a', 'b', 'c'])
+        ...
+        >>> st.line_chart(chart_data)
+
+        .. output::
+           https://share.streamlit.io/0.50.0-td2L/index.html?id=BdxXG3MmrVBfJyqS2R2ki8
+           height: 220px
+
+        """
+        vega_lite_chart_proto = VegaLiteChartProto()
+
+        chart = generate_chart("line", data, width, height)
+        marshall(vega_lite_chart_proto, chart, use_container_width)
+        last_index = last_index_for_melted_dataframes(data)
+
+        return dg._enqueue("line_chart", vega_lite_chart_proto, last_index=last_index)  # type: ignore
+
+    def area_chart(dg, data=None, width=0, height=0, use_container_width=True):
+        """Display a area chart.
+
+        This is just syntax-sugar around st.altair_chart. The main difference
+        is this command uses the data's own column and indices to figure out
+        the chart's spec. As a result this is easier to use for many "just plot
+        this" scenarios, while being less customizable.
+
+        Parameters
+        ----------
+        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, or dict
+            Data to be plotted.
+
+        width : int
+            The chart width in pixels. If 0, selects the width automatically.
+
+        height : int
+            The chart width in pixels. If 0, selects the height automatically.
+
+        use_container_width : bool
+            If True, set the chart width to the column width. This takes
+            precedence over the width argument.
+
+        Example
+        -------
+        >>> chart_data = pd.DataFrame(
+        ...     np.random.randn(20, 3),
+        ...     columns=['a', 'b', 'c'])
+        ...
+        >>> st.area_chart(chart_data)
+
+        .. output::
+           https://share.streamlit.io/0.50.0-td2L/index.html?id=Pp65STuFj65cJRDfhGh4Jt
+           height: 220px
+
+        """
+        vega_lite_chart_proto = VegaLiteChartProto()
+
+        chart = generate_chart("area", data, width, height)
+        marshall(vega_lite_chart_proto, chart, use_container_width)
+        last_index = last_index_for_melted_dataframes(data)
+
+        return dg._enqueue("area_chart", vega_lite_chart_proto, last_index=last_index)  # type: ignore
+
+    def bar_chart(dg, data=None, width=0, height=0, use_container_width=True):
+        """Display a bar chart.
+
+        This is just syntax-sugar around st.altair_chart. The main difference
+        is this command uses the data's own column and indices to figure out
+        the chart's spec. As a result this is easier to use for many "just plot
+        this" scenarios, while being less customizable.
+
+        Parameters
+        ----------
+        data : pandas.DataFrame, pandas.Styler, numpy.ndarray, Iterable, or dict
+            Data to be plotted.
+
+        width : int
+            The chart width in pixels. If 0, selects the width automatically.
+
+        height : int
+            The chart width in pixels. If 0, selects the height automatically.
+
+        use_container_width : bool
+            If True, set the chart width to the column width. This takes
+            precedence over the width argument.
+
+        Example
+        -------
+        >>> chart_data = pd.DataFrame(
+        ...     np.random.randn(50, 3),
+        ...     columns=["a", "b", "c"])
+        ...
+        >>> st.bar_chart(chart_data)
+
+        .. output::
+           https://share.streamlit.io/0.50.0-td2L/index.html?id=5U5bjR2b3jFwnJdDfSvuRk
+           height: 220px
+
+        """
+        vega_lite_chart_proto = VegaLiteChartProto()
+
+        chart = generate_chart("bar", data, width, height)
+        marshall(vega_lite_chart_proto, chart, use_container_width)
+        last_index = last_index_for_melted_dataframes(data)
+
+        return dg._enqueue("bar_chart", vega_lite_chart_proto, last_index=last_index)  # type: ignore
+
+    def altair_chart(dg, altair_chart, width=0, use_container_width=False):
+        """Display a chart using the Altair library.
+
+        Parameters
+        ----------
+        altair_chart : altair.vegalite.v2.api.Chart
+            The Altair chart object to display.
+
+        width : number
+            Deprecated. If != 0 (default), will show an alert.
+            From now on you should set the width directly in the Altair
+            spec. Please refer to the Altair documentation for details.
+
+        use_container_width : bool
+            If True, set the chart width to the column width. This takes
+            precedence over Altair's native `width` value.
+
+        Example
+        -------
+
+        >>> import pandas as pd
+        >>> import numpy as np
+        >>> import altair as alt
+        >>>
+        >>> df = pd.DataFrame(
+        ...     np.random.randn(200, 3),
+        ...     columns=['a', 'b', 'c'])
+        ...
+        >>> c = alt.Chart(df).mark_circle().encode(
+        ...     x='a', y='b', size='c', color='c', tooltip=['a', 'b', 'c'])
+        >>>
+        >>> st.altair_chart(c, use_container_width=True)
+
+        .. output::
+           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=8jmmXR8iKoZGV4kXaKGYV5
+           height: 200px
+
+        Examples of Altair charts can be found at
+        https://altair-viz.github.io/gallery/.
+
+        """
+        vega_lite_chart_proto = VegaLiteChartProto()
+
+        if width != 0:
+            import streamlit as st
+
+            st.warning(
+                "The `width` argument in `st.vega_lite_chart` is deprecated and will be removed on 2020-03-04. To set the width, you should instead use altair's native `width` argument as described at https://altair-viz.github.io/user_guide/generated/toplevel/altair.Chart.html"
+            )
+
+        marshall(
+            vega_lite_chart_proto,
+            altair_chart,
+            use_container_width=use_container_width,
+        )
+        return dg._enqueue("vega_lite_chart", vega_lite_chart_proto)  # type: ignore
 
 
 def _is_date_column(df, name):
